+++ c0622g1_auction/src/main/java/com/project/controller/users/UserRestController.java
@@ -147,20 +147,14 @@ public class UserRestController {
      * @return HttpStatus.OK if result is not empty
      * @return HttpStatus.NOT_FOUND if result is not empty
      */
-    @Autowired
-    private ILockAccountService lockAccountService;
+    @PostMapping("/add")
+    public ResponseEntity<?> addUser( @Validated @RequestBody FormAddUser formAddUser) {
 
-
-    @PostMapping("/create")
-    public ResponseEntity<?> addUser( @Validated @RequestBody AddUserDto addUserDto) {
-
-        AddressDto addressDto = new AddressDto(addUserDto.getDetailAddress(), addUserDto.getTown(),
-                addUserDto.getDistrict(), addUserDto.getCity(), addUserDto.getCountry());
-        AccountDto accountDto = new AccountDto(addUserDto.getUsername(), addUserDto.getPassword());
-
-        UserDto userDto = new UserDto(addUserDto.getFirstName(), addUserDto.getLastName(), addUserDto.getEmail(),
-                addUserDto.getPhone(), addUserDto.getPointDedication(), addUserDto.getBirthDay(),
-                addUserDto.getIdCard(), addUserDto.getAvatar(), addressDto, accountDto);
+       AddressDto addressDto= new AddressDto(formAddUser.getDetailAddress(),formAddUser.getTown(),formAddUser.getDistrict(),formAddUser.getCity(),formAddUser.getCountry());
+        AccountDto accountDto = new AccountDto(formAddUser.getUsername(), formAddUser.getPassword(),formAddUser.getStatusLock(),formAddUser.getDeleteStatus());
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getFirstName(), formAddUser.getLastName(), formAddUser.getEmail(),
+                formAddUser.getPhone(), formAddUser.getPointDedication(), formAddUser.getBirthDay(),
+                formAddUser.getIdCard(), formAddUser.getAvatar(), addressDto, accountDto);
 
         User user = new User();
         Address address = new Address();
@@ -168,18 +162,15 @@ public class UserRestController {
 
         BeanUtils.copyProperties(addressDto, address);
         BeanUtils.copyProperties(accountDto, account);
-        BeanUtils.copyProperties(userDto, user);
+        BeanUtils.copyProperties(addUserDto, user);
 
         Address addressATBC = addressService.saveAddress(address);
         Account accountABT = accountService.saveAccount(account);
-        accountABT.setStatusLock(true);
-        accountABT.setDeleteStatus(true);
-        accountABT.setPassword("12345678");
+
         userService.saveUser(user, addressATBC.getId(), accountABT.getId(), 4);
 
         return new ResponseEntity<>(HttpStatus.OK);
     }
-
 
     /**
      * Create by: VietNQ
