+++ c0622g1_auction/src/main/java/com/project/dto/user/AccountDto.java
@@ -4,6 +4,24 @@ public class AccountDto {
     private int id;
     private String username;
     private String password;
+    private Boolean statusLock;
+    private Boolean deleteStatus;
+
+
+    public AccountDto(String username, String password, Boolean statusLock , Boolean deleteStatus) {
+        this.username = username;
+        this.password = password;
+        this.statusLock = statusLock;
+        this.deleteStatus = deleteStatus;
+    }
+
+    public Boolean getStatusLock() {
+        return statusLock;
+    }
+
+    public void setStatusLock(Boolean statusLock) {
+        this.statusLock = statusLock;
+    }
 
     public AccountDto() {
     }
@@ -16,10 +34,6 @@ public class AccountDto {
         this.id = id;
     }
 
-    public AccountDto( String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
 
     public String getUsername() {
         return username;
@@ -36,4 +50,12 @@ public class AccountDto {
     public void setPassword(String password) {
         this.password = password;
     }
+    public Boolean getDeleteStatus() {
+        return deleteStatus;
+    }
+
+    public void setDeleteStatus(Boolean deleteStatus) {
+        this.deleteStatus = deleteStatus;
+    }
+
 }
