+++ c0622g1_auction/src/main/java/com/project/dto/user/AddUserDto.java
@@ -1,26 +1,50 @@
 package com.project.dto.user;
 
+
+import com.project.model.users.UserType;
+
+
 public class AddUserDto {
+
+    private Integer id;
     private String firstName;
+
     private String lastName;
     private String email;
+
     private String phone;
     private Double pointDedication;
+
     private String birthDay;
+
     private String idCard;
     private String avatar;
     private Boolean deleteStatus;
-    private String detailAddress;
-    private String town;
-    private String district;
-    private String city;
-    private String  country;
-    private String username;
-    private String password;
+    private AddressDto addressDto;
+    private AccountDto accountDto;
+    private UserType userType;
 
-    public AddUserDto() {
+    public AddUserDto(String firstName, String lastName, String email, String phone, Double pointDedication, String birthDay, String idCard, String avatar, AddressDto addressDto, AccountDto accountDto) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.email = email;
+        this.phone = phone;
+        this.pointDedication = pointDedication;
+        this.birthDay = birthDay;
+        this.idCard = idCard;
+        this.avatar = avatar;
+        this.addressDto = addressDto;
+        this.accountDto = accountDto;
     }
 
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
 
     public String getFirstName() {
         return firstName;
@@ -94,59 +118,21 @@ public class AddUserDto {
         this.deleteStatus = deleteStatus;
     }
 
-    public String getDetailAddress() {
-        return detailAddress;
+    public AddressDto getAddressDto() {
+        return addressDto;
     }
 
-    public void setDetailAddress(String detailAddress) {
-        this.detailAddress = detailAddress;
+    public void setAddressDto(AddressDto addressDto) {
+        this.addressDto = addressDto;
     }
 
-    public String getTown() {
-        return town;
+    public AccountDto getAccountDto() {
+        return accountDto;
     }
 
-    public void setTown(String town) {
-        this.town = town;
+    public void setAccountDto(AccountDto accountDto) {
+        this.accountDto = accountDto;
     }
 
-    public String getDistrict() {
-        return district;
-    }
 
-    public void setDistrict(String district) {
-        this.district = district;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
 }
