+++ c0622g1_auction/src/main/java/com/project/dto/user/AddressDto.java
@@ -1,6 +1,5 @@
 package com.project.dto.user;
 
-import com.project.model.users.User;
 
 public class AddressDto {
     private Integer id;
@@ -9,7 +8,7 @@ public class AddressDto {
     private String district;
     private String city;
     private String  country;
-    private User user;
+
 
     public AddressDto() {
     }
@@ -68,13 +67,5 @@ public class AddressDto {
 
     public void setCountry(String country) {
         this.country = country;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
     }
 }
