+++ c0622g1_auction/src/main/java/com/project/dto/user/FormAddUser.java
@@ -1,0 +1,217 @@
+package com.project.dto.user;
+
+import com.project.model.users.User;
+import org.springframework.validation.Errors;
+import org.springframework.validation.Validator;
+
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+import java.util.List;
+
+public class FormAddUser implements Validator {
+
+    private Integer id;
+    @NotEmpty(message = "Họ không được để trống")
+    private String firstName;
+    @NotEmpty(message = "Tên không được để trống")
+    private String lastName;
+    @NotEmpty(message = "Email không được để trống")
+    private String email;
+    @NotEmpty(message = "Số điện thoại không được để trống")
+    private String phone;
+    @NotNull(message = "Điểm cống hiến không được bỏ trống")
+    @Min(value = 0, message = "Điểm cống hiến không được bé thua 0")
+    @Max(value = 500, message = "Điểm cống hiến không được lớn hơn 500")
+    private Double pointDedication;
+    @NotEmpty(message = "Ngày sinh không được để trống")
+    private String birthDay;
+    @NotEmpty(message = "CMND không được để trống")
+    private String idCard;
+    @NotEmpty(message = "Avatar được để trống")
+    private String avatar;
+    private String username;
+    private String password;
+    @NotEmpty(message = "Không được để trống")
+    private String detailAddress;
+    @NotEmpty(message = "Không được để trống")
+    private String town;
+    @NotEmpty(message = "Không được để trống")
+    private String district;
+    @NotEmpty(message ="Không được để trống")
+    private String city;
+    @NotEmpty(message ="Không được để trống")
+    private String  country;
+    private Boolean statusLock;
+    private Boolean deleteStatus;
+
+
+
+    public Boolean getStatusLock() {
+        return statusLock;
+    }
+
+    public void setStatusLock(Boolean statusLock) {
+        this.statusLock = statusLock;
+    }
+
+    private List<String> emailList;
+
+    public List<String> getEmailList() {
+        return emailList;
+    }
+
+    public void setEmailList(List<String> emailList) {
+        this.emailList = emailList;
+    }
+
+    public FormAddUser() {
+    }
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+
+    public Double getPointDedication() {
+        return pointDedication;
+    }
+
+    public void setPointDedication(Double pointDedication) {
+        this.pointDedication = pointDedication;
+    }
+
+    public String getBirthDay() {
+        return birthDay;
+    }
+
+    public void setBirthDay(String birthDay) {
+        this.birthDay = birthDay;
+    }
+
+    public String getIdCard() {
+        return idCard;
+    }
+
+    public void setIdCard(String idCard) {
+        this.idCard = idCard;
+    }
+
+    public String getAvatar() {
+        return avatar;
+    }
+
+    public void setAvatar(String avatar) {
+        this.avatar = avatar;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getDetailAddress() {
+        return detailAddress;
+    }
+
+    public void setDetailAddress(String detailAddress) {
+        this.detailAddress = detailAddress;
+    }
+
+    public String getTown() {
+        return town;
+    }
+
+    public void setTown(String town) {
+        this.town = town;
+    }
+
+    public String getDistrict() {
+        return district;
+    }
+
+    public void setDistrict(String district) {
+        this.district = district;
+    }
+
+    public String getCity() {
+        return city;
+    }
+
+    public void setCity(String city) {
+        this.city = city;
+    }
+
+    public String getCountry() {
+        return country;
+    }
+
+    public void setCountry(String country) {
+        this.country = country;
+    }
+
+
+    @Override
+    public boolean supports(Class<?> clazz) {
+        return false;
+    }
+
+    @Override
+    public void validate(Object target, Errors errors) {
+
+    }
+    public Boolean getDeleteStatus() {
+        return deleteStatus;
+    }
+
+    public void setDeleteStatus(Boolean deleteStatus) {
+        this.deleteStatus = deleteStatus;
+    }
+}
