+++ c0622g1_auction/src/main/java/com/project/repository/account/ILockAccountRepository.java
@@ -2,10 +2,12 @@ package com.project.repository.account;
 
 import com.project.model.account.LockAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 @Transactional
 public interface ILockAccountRepository extends JpaRepository<LockAccount, Integer> {
+
 }
