+++ c0622g1_auction/src/main/java/com/project/service/account/impl/LockAccountService.java
@@ -1,11 +1,10 @@
 package com.project.service.account.impl;
 
-import com.project.repository.account.ILockAccountRepository;
 import com.project.service.account.ILockAccountService;
 import org.springframework.stereotype.Service;
 
 
 @Service
 public class LockAccountService implements ILockAccountService {
-    private ILockAccountRepository lockAccountRepository;
+
 }
