+++ c0622g1_auction/src/test/java/com/project/controllor/AuctionControllerAddUser.java
@@ -4,7 +4,7 @@ package com.project.controllor;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.dto.user.AccountDto;
 import com.project.dto.user.AddressDto;
-import com.project.dto.user.UserDto;
+import com.project.dto.user.AddUserDto;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -32,22 +32,22 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_fisrt_name_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setLastName("Nam");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setLastName("Nam");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -60,23 +60,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_fisrt_name_14() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("");
-        userDto.setLastName("Nam");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("");
+        addUserDto.setLastName("Nam");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -89,23 +89,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_fisrt_name_15() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("vietaaasa");
-        userDto.setLastName("Nam");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("vietaaasa");
+        addUserDto.setLastName("Nam");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -118,23 +118,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_fisrt_name_16() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("vietaaasxxxxxxxxxxxxa");
-        userDto.setLastName("Nam");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("vietaaasxxxxxxxxxxxxa");
+        addUserDto.setLastName("Nam");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -147,23 +147,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_fisrt_name_17() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("v");
-        userDto.setLastName("Nam");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("v");
+        addUserDto.setLastName("Nam");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -177,22 +177,22 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_gmail_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -205,23 +205,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_gmail_14() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("Viet");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("Viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -234,23 +234,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_gmail_15() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("Namasasasas");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("agmailcom");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("Namasasasas");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("agmailcom");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -263,23 +263,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_gmail_16() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -292,23 +292,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_gmail_17() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("Viet");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("123456789");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("Viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -322,22 +322,22 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_phone_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -350,23 +350,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_phone_14() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -379,23 +379,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_phone_15() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("Namasasasas");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("dadaaw222");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("Namasasasas");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("dadaaw222");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -408,23 +408,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_phone_16() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("1234567894444444444444444444444444444444444444");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("1234567894444444444444444444444444444444444444");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -437,23 +437,23 @@ public class AuctionControllerAddUser {
      */
     @Test
     public void addUser_phone_17() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setAvatar("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setPhone("1");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setAvatar("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setPhone("1");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -467,22 +467,22 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_avatar_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -496,23 +496,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_avatar_14() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -526,23 +526,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_avatar_15() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("1231e121");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("1231e121");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -556,23 +556,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_avatar_16() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("wwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwww");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("wwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwwww");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -586,23 +586,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_avatar_17() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("a");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("1234");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("a");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("1234");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -616,22 +616,22 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_idCard_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -645,23 +645,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_idCard_14() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -675,23 +675,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_idCard_15() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("qqqqqqqqqqqqqqqq");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("qqqqqqqqqqqqqqqq");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -705,23 +705,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_idCard_16() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setIdCard("1234888888888888888888888888888888");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("1234888888888888888888888888888888");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -735,23 +735,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_idCard_17() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("1");
-        userDto.setBirthDay("12/12/2021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("1");
+        addUserDto.setBirthDay("12/12/2021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -765,22 +765,22 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_birthDay_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setIdCard("12345678");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -794,23 +794,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_birthDay_14() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -824,23 +824,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_birthday_15() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("adadadadadadada");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("adadadadadadada");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -854,23 +854,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_birthDay_16() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("121/112/22021");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("121/112/22021");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -884,23 +884,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_birthDay_17() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("1/1/202");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("1/1/202");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -914,21 +914,21 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_pointDedication_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setIdCard("12345678");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -942,23 +942,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_pointDedication_14() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(0.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("12/12/2022");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(0.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("12/12/2022");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -972,23 +972,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_pointDedication_15() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(010.00);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("12/12/1022");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(010.00);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("12/12/1022");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -1002,23 +1002,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_pointDedication_16() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(100000000000000000000.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("12/12/2202");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(100000000000000000000.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("12/12/2202");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -1032,23 +1032,23 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_pointDedication_17() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setAvatar("viet");
-        userDto.setPointDedication(0.1);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("11/11/2022");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setAvatar("viet");
+        addUserDto.setPointDedication(0.1);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("11/11/2022");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -1062,20 +1062,20 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_AdressDto_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setPointDedication(10.0);
-        userDto.setLastName("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setIdCard("12345678");
-        userDto.setAccountDto(accountDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setLastName("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setAccountDto(accountDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -1089,20 +1089,20 @@ public class AuctionControllerAddUser {
 
     @Test
     public void addUser_accountDto_13() throws Exception {
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setPointDedication(10.0);
-        userDto.setLastName("viet");
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setIdCard("12345678");
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setLastName("viet");
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setAddressDto(addressDto);
 
         this.mockMvc.perform(MockMvcRequestBuilders
                         .post("/api/user/v1/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is4xxClientError());
@@ -1116,23 +1116,23 @@ public class AuctionControllerAddUser {
     @Test
     public void createStudent_18() throws Exception {
 
-        UserDto userDto = new UserDto();
+        AddUserDto addUserDto = new AddUserDto(formAddUser.getDistrict(), formAddUser.getTown(), formAddUser.getDistrict(), formAddUser.getCity(), formAddUser.getCountry());
         AddressDto addressDto = new AddressDto();
         AccountDto accountDto = new AccountDto();
-        userDto.setFirstName("Nguyen");
-        userDto.setLastName("viet");
-        userDto.setPointDedication(10.0);
-        userDto.setEmail("a@gmail.com");
-        userDto.setPhone("123456789");
-        userDto.setAvatar("viet");
-        userDto.setIdCard("12345678");
-        userDto.setBirthDay("12/12/2202");
-        userDto.setAccountDto(accountDto);
-        userDto.setAddressDto(addressDto);
+        addUserDto.setFirstName("Nguyen");
+        addUserDto.setLastName("viet");
+        addUserDto.setPointDedication(10.0);
+        addUserDto.setEmail("a@gmail.com");
+        addUserDto.setPhone("123456789");
+        addUserDto.setAvatar("viet");
+        addUserDto.setIdCard("12345678");
+        addUserDto.setBirthDay("12/12/2202");
+        addUserDto.setAccountDto(accountDto);
+        addUserDto.setAddressDto(addressDto);
         this.mockMvc
                 .perform(MockMvcRequestBuilders
                         .post("/studentRest/create")
-                        .content(this.objectMapper.writeValueAsString(userDto))
+                        .content(this.objectMapper.writeValueAsString(addUserDto))
                         .contentType(MediaType.APPLICATION_JSON_VALUE))
                 .andDo(print())
                 .andExpect(status().is2xxSuccessful());
